First implementation of the basic methods finished
