--- client_secret.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"172263236622-5qs638225pbfmqvnr4i35k2hnp8p1vir.apps.googleusercontent.com","project_id":"gsheets-lib","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"L95F2c9qJXTB2M1p-DTOuflw","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
